+++ packages/contentful--app-scripts/mocha-setup.js
@@ -1,7 +1,3 @@
 'use strict';
 
 process.env.NODE_ENV = 'test';
-
-require('ts-node').register({
-  transpileOnly: true,
-});