+++ packages/contentful--app-scripts/src/types.ts
@@ -58,6 +58,11 @@ export interface OpenSettingsOptions {
   definitionId?: string;
 }
 
+export interface InstallOptions {
+  definitionId?: string;
+  host?: string;
+}
+
 export interface UploadOptions {
   organizationId?: string;
   definitionId?: string;